chore: moved ModuleVersionData into own file
