+++ JotunnTests/CompatibilityLevelTest.cs
@@ -10,8 +10,8 @@ namespace Jotunn.Utils
         [TestMethod]
         public void BothOnlyJotunn()
         {
-            var clientVersionData = new ModCompatibility.ModuleVersionData(new System.Version(1, 0, 0), new List<Tuple<string, System.Version, CompatibilityLevel, VersionStrictness>>());
-            var serverVersionData = new ModCompatibility.ModuleVersionData(new System.Version(1, 0, 0), new List<Tuple<string, System.Version, CompatibilityLevel, VersionStrictness>>());
+            var clientVersionData = new ModuleVersionData(new System.Version(1, 0, 0), new List<Tuple<string, System.Version, CompatibilityLevel, VersionStrictness>>());
+            var serverVersionData = new ModuleVersionData(new System.Version(1, 0, 0), new List<Tuple<string, System.Version, CompatibilityLevel, VersionStrictness>>());
 
             Assert.IsTrue(ModCompatibility.CompareVersionData(serverVersionData, clientVersionData));
         }
@@ -23,9 +23,9 @@ namespace Jotunn.Utils
             {
                 new Tuple<string, System.Version, CompatibilityLevel, VersionStrictness>("TestMod", new System.Version(1, 0, 0), CompatibilityLevel.EveryoneMustHaveMod, VersionStrictness.Minor)
             };
-            var clientVersionData = new ModCompatibility.ModuleVersionData(new System.Version(1, 0, 0), clientMods);
+            var clientVersionData = new ModuleVersionData(new System.Version(1, 0, 0), clientMods);
             var serverMods = new List<Tuple<string, System.Version, CompatibilityLevel, VersionStrictness>>();
-            var serverVersionData = new ModCompatibility.ModuleVersionData(new System.Version(1, 0, 0), serverMods);
+            var serverVersionData = new ModuleVersionData(new System.Version(1, 0, 0), serverMods);
 
             Assert.IsFalse(ModCompatibility.CompareVersionData(serverVersionData, clientVersionData));
         }
@@ -34,12 +34,12 @@ namespace Jotunn.Utils
         public void ServerHasModButClientDoesntNeedIt()
         {
             var clientMods = new List<Tuple<string, System.Version, CompatibilityLevel, VersionStrictness>>();
-            var clientVersionData = new ModCompatibility.ModuleVersionData(new System.Version(1, 0, 0), clientMods);
+            var clientVersionData = new ModuleVersionData(new System.Version(1, 0, 0), clientMods);
             var serverMods = new List<Tuple<string, System.Version, CompatibilityLevel, VersionStrictness>>
             {
                 new Tuple<string, System.Version, CompatibilityLevel, VersionStrictness>("TestMod", new System.Version(1, 0, 0), CompatibilityLevel.ServerMustHaveMod, VersionStrictness.Minor)
             };
-            var serverVersionData = new ModCompatibility.ModuleVersionData(new System.Version(1, 0, 0), serverMods);
+            var serverVersionData = new ModuleVersionData(new System.Version(1, 0, 0), serverMods);
 
             Assert.IsTrue(ModCompatibility.CompareVersionData(serverVersionData, clientVersionData));
         }
